Tagging the FF1 branch of hdf5_ff

+++ hl/tools/gif2h5/decompress.c
@@ -90,7 +90,7 @@ ReadCode(void)
         RawCode += (0x10000 * Raster[ByteOffset + 2]);
 
     RawCode >>= (BitOffset % 8);
-    BitOffset += (int)CodeSize;
+    BitOffset += CodeSize;
     return (RawCode & ReadMask);
 }
